requirements.txt
asgiref==3.10.0
certifi==2025.10.5
charset-normalizer==3.4.4
Django==5.1.2
django-redis==5.4.0
djangorestframework==3.15.2
drf-spectacular==0.27.2
idna==3.11
psycopg2-binary==2.9.9
redis==6.4.0
requests==2.32.3
setuptools==80.9.0
sqlparse==0.5.3
typing_extensions==4.15.0
urllib3==2.5.0
wheel==0.45.1
